Move game role functions out of helper
